backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Supabase
supabase==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data fetching
yfinance==0.2.28
python-binance==1.0.19
requests==2.31.0

# Trading and backtesting
backtrader==1.9.78.123
pandas==2.1.4
numpy==1.25.2

# Risk analysis
scipy==1.11.4
scikit-learn==1.3.2

# Paper trading
alpaca-py==0.20.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

